+++ plugin_ide.ui/src/LANG_PROJECT_ID/ide/ui/editor/LANGUAGE_SourceViewerConfiguration.java
@@ -13,6 +13,7 @@ package LANG_PROJECT_ID.ide.ui.editor;
 import static melnorme.utilbox.core.CoreUtil.array;
 import melnorme.lang.ide.ui.LangUIPlugin_Actual;
 import melnorme.lang.ide.ui.TextSettings_Actual.LangPartitionTypes;
+import melnorme.lang.ide.ui.editor.AbstractLangEditor;
 import melnorme.lang.ide.ui.text.AbstractLangSourceViewerConfiguration;
 
 import org.eclipse.cdt.ui.text.IColorManager;
@@ -27,7 +28,7 @@ import LANG_PROJECT_ID.ide.ui.text.LANGUAGE_ColorPreferences;
 public class LANGUAGE_SourceViewerConfiguration extends AbstractLangSourceViewerConfiguration {
 	
 	public LANGUAGE_SourceViewerConfiguration(IPreferenceStore preferenceStore, IColorManager colorManager,
-			LANGUAGE_Editor editor) {
+			AbstractLangEditor editor) {
 		super(preferenceStore, colorManager, editor);
 	}