+++ documentation/ChangeLog.md
@@ -2,7 +2,9 @@
 
 ### LANG_IDE_NAME 0.0.x
  * Added New Project wizard.
-  
+ 
   ▶ Recommended/tested CDT version: 8.5.0
+  
+ * Added: Source menu with shift rigth/left operations.